Implemented customisable base actions, and refactor

+++ lib/railgun/resources_controller/actions.rb
@@ -4,8 +4,11 @@ module Railgun
 			def index(options={}, &block)
 				super(options) do |format|
 					@current_scope = current_scope_key
+
+					# TODO: should be elsewhere..
 					set_title(railgun_resource.name.pluralize)
-					add_action_button(:default, "Add New", [:new, railgun_resource.to_sym], :type => "info") 
+					add_action_button(:default, "Add New", [:new, railgun_resource.to_sym], :type => "info") if railgun_resource.actions.include?(:new)
+
 					instance_exec format, &block if block_given?
 					format.html { render_railgun railgun_template("resources/index") }
 				end
@@ -15,9 +18,9 @@ module Railgun
 				super(options) do |format|
 					add_crumb(:title => resource.send(railgun_resource.name_column), :path => [resource])
 					set_title("View "+railgun_resource.name)
-					add_action_button(:default, "Edit", [:edit, resource], :type => "info")
+					add_action_button(:default, "Edit", [:edit, resource], :type => "info") if railgun_resource.actions.include?(:edit)
 					add_action_button(:destroy, "Delete", resource, 
-						:type => "danger", :method => :delete, :confirm => "Are you sure you want to delete this record?") 
+						:type => "danger", :method => :delete, :confirm => "Are you sure you want to delete this record?")  if railgun_resource.actions.include?(:destroy)
 					instance_exec format, &block if block_given?
 					format.html { render_railgun railgun_template("resources/show") }
 				end
